requirements.txt
amqp==2.3.2
aniso8601==3.0.2
argon2-cffi==18.1.0
asn1crypto==0.24.0
bcrypt==3.1.4
beautifulsoup4==4.6.3
billiard==3.5.0.4
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.1
django-adminlte2==0.4.1
django-allauth==0.37.1
django-cors-headers==2.4.0
django-extensions==2.1.0
django-heroku==0.3.1
django-rest-auth==0.9.3
django-tawkto==0.1
django-watson==1.5.2
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
drf-nested-routers==0.90.2
Flask==1.0.2
Flask-Cors==3.0.6
flask-restplus==0.11.0
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
kombu==4.2.1
MarkupSafe==1.0
mccabe==0.6.1
oauthlib==2.1.0
pendulum==2.0.3
Pillow==5.2.0
psycopg2==2.7.5
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
PyJWT==1.6.4
pymongo==3.7.1
pyOpenSSL==18.0.0
python-dateutil==2.7.3
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.5
pytzdata==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
simplejson==3.16.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
text-unidecode==1.2
urllib3==1.23
vine==1.1.4
Werkzeug==0.14.1
whitenoise==4.0
